A7: extract @gb-rail/gtfs-output
FileOutput and the zip command move into libs/gtfs-output, along with
processSpawnResult, which nothing else used. The GTFSOutput interface itself
stays in libs/gtfs - see A6.

OutputGTFSZipCommand moves unchanged, setTimeout race and all. Fixing that is
ticket B5, and doing it in the same commit as a move would make the diff
unreviewable.

GTFS output verified byte-identical to the pre-restructure build.
