+++ libs/gtfs-output/package.json
@@ -1,0 +1,35 @@
+{
+  "name": "@gb-rail/gtfs-output",
+  "version": "0.0.0",
+  "description": "Writers for a GTFS feed: a directory of text files, or a zip",
+  "license": "GPL-3.0",
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/planarnetwork/dtd2mysql.git",
+    "directory": "libs/gtfs-output"
+  },
+  "main": "./dist/index.js",
+  "types": "./dist/index.d.ts",
+  "exports": {
+    ".": {
+      "types": "./dist/index.d.ts",
+      "default": "./dist/index.js"
+    }
+  },
+  "files": [
+    "dist"
+  ],
+  "publishConfig": {
+    "access": "public"
+  },
+  "scripts": {
+    "build": "tsc -b"
+  },
+  "dependencies": {
+    "@gb-rail/gtfs": "workspace:^",
+    "csv-write-stream": "2.0.0"
+  },
+  "devDependencies": {
+    "@types/csv-write-stream": "2.0.3"
+  }
+}
