+++ libs/gtfs-output/src/FileOutput.ts
