+++ libs/gtfs-output/src/index.ts
@@ -1,0 +1,2 @@
+export {FileOutput} from "./FileOutput";
+export {OutputGTFSZipCommand} from "./OutputGTFSZipCommand";
